feat(ui): aplica visual do prototipo em Detalhes e Meu Deck, e corrige notas decimais (closes #23, closes #24)

+++ internal/entries/models.go
@@ -6,6 +6,6 @@ type MediaEntry struct {
 	MalID    int    `json:"mal_id"`
 	Tipo     string `json:"tipo"`
 	Status   string `json:"status"`
-	Nota     *int   `json:"nota,omitempty"`
+	Nota     *float64   `json:"nota,omitempty"`
 	Anotacao string `json:"anotacao,omitempty"`
 }
